requirements.txt
botorch==0.9.2
control==0.9.4
gpytorch==1.11
matplotlib
numpy
pyro_ppl
slycot==0.5.4
torch==2.5.0
tqdm==4.67.1
